+++ consumers/CHANGELOG.md
@@ -1,3 +1,6 @@
+# consumers-2.3.5.0 (2026-??-??)
+* Add `mapConsumer` to ` Database.PostgreSQL.Consumers.Config`.
+
 # consumers-2.3.4.0 (2025-11-27)
 * Compatibility with `hpqtypes` >= 1.13.0.0.